speed up and fixing

+++ sep/extract.py
@@ -18,7 +18,7 @@ def extract_to_images(data_loader: Loader, data_saver: Saver,
         print(f"There are {len(data_loader)} images to process.")
 
     data_saver.set_output(output_root, data_loader)
-    if remove_existing:
+    if remove_existing and os.path.exists(output_root):
         shutil.rmtree(output_root)
     os.makedirs(output_root)
